Split off more stuff from utils.hpp
